Split the multer upload logic into a new file and referenced from the about-your-income routes file.
