+++ alpha/external-web/app/routes/about-your-income.js
@@ -1,28 +1,6 @@
 var router = require('../routes')
 var client = require('../services/eligibility-client')
-
-const maxFileSize = 5242880 // 5MB in Bytes.
-const allowedFileTypes = ['image/png', 'image/jpeg', 'application/pdf']
-
-function fileFilter (req, file, callback) {
-  var mimeType = file.mimetype
-
-  if (!allowedFileTypes.includes(mimeType)) {
-    var error = 'Uploaded file was not an image.'
-    req.fileValidationError = error
-    return callback(null, false, new Error(error))
-  }
-  callback(null, true)
-}
-
-// Require file upload library.
-var multer = require('multer')({
-  dest: 'eligibility-uploads/',
-  limits: {
-    fileSize: maxFileSize
-  },
-  fileFilter: fileFilter
-})
+var upload = require('../services/upload')
 
 const claimantsCollection = 'claimants'
 
@@ -37,7 +15,7 @@ router.get('/about-your-income/:claimant_id', function (request, response, next)
   next()
 })
 
-router.post('/about-your-income/:claimant_id', multer.single('evidence'), function (request, response, next) {
+router.post('/about-your-income/:claimant_id', upload.single('evidence'), function (request, response, next) {
   var id = request.params.claimant_id
   if (!request.file) {
     response.status(500).render('error', { error: 'Failed to update claimant with id: ' + id + '. No file was uploaded.' })
